Use tpm's expected environment variable for storing resurrect files
This fixes tmux-resurrect for cases where the user has setup tpm to
install tmux plugins to a directory that is non-default (i.e:
"$XDG_CONFIG_HOME/tmux" instead of "$HOME/.tmux").

This allows the plugin to be more flexible based on the user's
configuration but still supports the default... by default.